refactor: extract health status generation into a reusable service #279
